requirements-api.txt
ultralytics==8.0.200
opencv-python-headless==4.8.1.78
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
Pillow==10.1.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
python-dotenv==1.0.0
